Finiquitada ida y casi toa la vuelta

+++ Assets/Scripts/_Scripts Prólogo/Plataforma tiempo/Scr_PlataformaTiempo.cs
@@ -8,7 +8,7 @@ public class Scr_PlataformaTiempo : MonoBehaviour
     bool doBreak;
     
     [SerializeField]
-    int timeToBreak;
+    float timeToBreak;
 
     [SerializeField]
     int timeToSpawn;
